+++ src/app/core/material/material.module.ts
@@ -4,12 +4,16 @@ import { MatCardModule } from '@angular/material/card';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatIconModule } from '@angular/material/icon';
 import { MatBadgeModule } from '@angular/material/badge';
+import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
 import { MatSelectModule } from '@angular/material/select';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatDividerModule } from '@angular/material/divider';
+import { MatStepperModule } from '@angular/material/stepper';
 import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatExpansionModule } from '@angular/material/expansion';
 import { MatBottomSheetModule } from '@angular/material';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 
 const materialModules = [
   MatListModule,
@@ -17,12 +21,16 @@ const materialModules = [
   MatTabsModule,
   MatIconModule,
   MatBadgeModule,
+  MatInputModule,
   MatButtonModule,
   MatSelectModule,
   MatToolbarModule,
   MatDividerModule,
+  MatStepperModule,
   MatCheckboxModule,
-  MatBottomSheetModule
+  MatExpansionModule,
+  MatBottomSheetModule,
+  MatSlideToggleModule
 ];
 
 @NgModule({