+++ src/app/modules/order/models/order.models.ts
@@ -1,12 +1,14 @@
 import { Product } from '../../products/models/product.model';
+import { ShippingDetails } from './shipping-details.models';
 
 export class Order {
   constructor(
-    public id: number,
-    public items: Product[],
-    public price: number,
-    public address: string,
-    public date: number,
-    public completed: boolean = false
+    public products: Product[],
+    public firstName: string,
+    public lastName: string,
+    public email: string,
+    public phones: string[],
+    public shippingDetails: ShippingDetails,
+    public id?: number
   ) {}
 }