+++ src/app/modules/order/models/shipping-details.models.ts
@@ -1,0 +1,10 @@
+export class ShippingDetails {
+  constructor(
+    public city: string,
+    public address: string,
+    public firstName: string,
+    public lastName: string,
+    public email: string,
+    public phones: string[]
+  ) {}
+}